#25 스트리머 유저 구독 비용 long -> int

+++ module-user/src/main/kotlin/com/example/user/application/port/in/CreateStreamerUserUseCase.kt
@@ -12,7 +12,7 @@ interface CreateStreamerUserUseCase {
 data class CreateStreamerUserCommand(
     val normalUser: NormalUser,
     val streamerUserNickname: String,
-    val subscriptionCost: Long
+    val subscriptionCost: Int
 ) {
     val id: UUID
         get() {
