+++ module-user/src/main/kotlin/com/example/user/domain/model/StreamerUser.kt
@@ -11,7 +11,7 @@ class StreamerUser(
     password: String,
     nickname: String,
     var streamerNickname: String,
-    var subscriptionCost: Long
+    var subscriptionCost: Int
 ) : User.Editor {
     override var password: String = ""
     override var status: UserStatus = UserStatus.REGISTERED
